pddl: add enter-field and mark-visited actions
Add the required enter-field action that is required for proper driving
into the field. The action mapping shows chaining of regexes nicely.
Even though that without parallel dispatching it is a little pointless
to wait before entering the field. What would actually be required is
something to ensure that the entered robot leaves the spot before the
next robot enters. Something for another iteration...
Make the move operator more generic by removing the visited effect. This
is now handled explicitly by the always succeeding mark-visited action
which makes sure to assert to the visited effect.