pddl: fix conditions/effects, marking non-durative, congestion fix
Needed to convert (at start (and (p1) (p2))) to (and (at start (p1)) (at
start (pe))). Seems that this trivial conversion is not automatically
handled. Could have known looking at the 2.1 paper...

Make mark-visited non-durative, it's not.

Fix the congestion that could happen in a multi-robot setup on entering
the field (if they all entered before the robots moved out of the way).
Now we explicitly mark free locations and essentially handle location
locking to make sure the robot moves out of the way before the next
robot exits.