pddl: allow prepare of CS, DS, and RS only when at machine
We parametrize the prepare action by a robot and require that the robot
be standing at the machine. That severaly reduces the number of actions
applicable in many situations leading to a drastically improved planning
performance. For the BS we still allow to do this any time because of
our constraint movement operators. Adding in a specialized BS move
operator to allow for both sides and preparing while at the machine has
a detrimental effect on planning time.

Interestingly this is also beneficial for the RCLL rules 2017 where the
preparation of a machine may occur only at most 30 seconds before using
it.