domain: remove specialized move operator for wp-put-slide-cc
The additional move operator lead to much longer planning times. With
the slight reformulation from the previous post (that we can move to a
machine when it is IDLE) we do not need the extra operator (which
essentially differed in that it allowed to go to a machine no matter its
current state).